fix(schema): resolve 3 auditor mismatches that aborted startup
Root cause: validateSchema() correctly found 3 diffs between the Drizzle
schema and the live DB, threw, and process.exit(1) fired — Railway returned 502.

Diff 1 — media_asset_cache.durationSec type mismatch
  schema declared int(), migration 0030 created the column as FLOAT.
  Fix: change schema.ts to float('durationSec') — no migration needed
  (DB already has the correct type; fractional seconds make sense for duration).

Diff 2 — editorial_reviews.overallScore type mismatch
  schema declares int(), migration 0033 created TINYINT UNSIGNED.
  Fix: migration 0034 widens the column to INT to match the schema.

Diff 3 — media_archives.createdByUserId FK missing
  schema declares .references(() => users.id) but migration 0012 never
  added the FOREIGN KEY constraint to the DB.
  Fix: migration 0034 adds it idempotently via INFORMATION_SCHEMA check.

+++ drizzle/schema.ts
@@ -1,4 +1,4 @@
-import { index, int, json, longtext, mysqlEnum, mysqlTable, text, timestamp, varchar } from "drizzle-orm/mysql-core";
+import { float, index, int, json, longtext, mysqlEnum, mysqlTable, text, timestamp, varchar } from "drizzle-orm/mysql-core";
 
 /**
  * Core user table backing auth flow.
@@ -415,8 +415,8 @@ export const mediaAssetCache = mysqlTable("media_asset_cache", {
   r2Key: varchar("r2Key", { length: 512 }).notNull(),
   contentType: varchar("contentType", { length: 64 }).notNull().default("application/octet-stream"),
   fileSizeBytes: int("fileSizeBytes").notNull().default(0),
-  /** Duration in seconds for video assets; null for images. */
-  durationSec: int("durationSec"),
+  /** Duration in seconds for video assets; null for images. Stored as FLOAT (fractional seconds). */
+  durationSec: float("durationSec"),
   /** Bump to invalidate all entries (e.g. if encoding settings change). */
   cacheVersion: varchar("cacheVersion", { length: 32 }).notNull().default("1"),
   hitCount: int("hitCount").notNull().default(0),
